feat: github integration
This commit sets a foundation or core for further development. It
implements all the grunt work of integrating with GitHub, namely:

* OAuth2 integration, basically login with GitHub
* A webclient which takes care of authorizing requests based on their
  scope (app or installation type requests)
* Webhook signature validation, this enables the use of secrets for
  webhooks
* A webhook to keep track of installation events with support for
  "created" and "deleted" events.

It also provides a crude yet working UI, created with Thymeleaf
templates and tailwindcss, with the build process integrated into Maven.

Finally, SonarCube was implemented into the CI pipeline which also
involves running tests on every PR.