feat: create repository page
This commit sets a change in direction for development, by basically
making the app a one controller application. This controller handles
the same URL as Github does for repos (e.g. /owner/project) which means
it can be accessed by swapping the domain (github.com) for felf's
domain.

This means there are no dashboards or other pages/controllers, besides
a login and index page. After all the UI is really there to create an
access token that can be used in the repo's CI to collect statistics.
The real stuff happens on Github.com with automatic bot comments on PRs
and good looking badges.