feat: create repository installations
The option to create projects (repository installations) has been
added. Projects contain all the measurements for a given repository
and the access tokens linked to it.

Projects are linked to an account installation, when no projects exist,
the user is prompted to add one. The addition process displays a modal
with all the repositories that have no linked projects, upon clicking a
new project is created and the user redirected to it.