ref: use base64 encoded pem file contents
The PEM file used to sign JSON Web Tokens was being previously read from
a file on disk, this makes it harder to use modern cloud deployment
technologies.

Therefore instead of reading from a file, the PEM is now provided
through a base64 encoded representation of it, this makes it possible to
pass it to the server as an environment variable.