feat: add columns to track analysis comments state
This is the first step in implementing a scheduled task to publish
analysis reports on pull requests.

NOOP is the default and will be used when the job is executing on a push comment.
TODO is applied upon receiving a new analysis from a PR. FAIL states there was a failed
attempt to publish an analysis report on the PR.

The previous comment id is also stored in order to delete it before
posting a new one.