ref(repo pages): separate not found view
In order to reduce the complexity of templates, the not found page has
been split into its own file. This makes the templates easier to read
and it moves logic into code and away from templates. As last benefit,
it is more efficient since the controller's method performs an early
exit if no project is found.